CB-3440 [BlackBerry10] Proxy based implementation
This is a re-write of the file plugin for BB10. Rather than clobbering the
window objects this utilizes the exec proxy to handle requests from common
JS in the content webview. All operations are handled by the webkit HTML5
file system API. The setSandbox function in index.js allows the file system
to be toggled between full device filesystem access and the app's local
sandbox.