CB-6525 android, ios: Allow file: URLs in all APIs. Fixes FileTransfer.download not being called.
Root problem of downlaod was that the target URL was not being properly
resolved to an Entry on the native side. This wasn't caught by
mobilespec due to the requestFileSystem() using cdvfile: URLs still,
while resolveLocalFileSystemURL(entry.toURL()), would yield a file: URL.

This commit normalizes fs.root DirectoryEntries so that they are file://
as well, and initializes them on start-up with a new "getAllFileSystems"
exec() call.

It also aims to ensure that JSON entries always have nativeURL set.