CB-6923 Adding support to handle relative paths
Added support to resolve relative paths.
When the path received is relative, it will combine and obtain a full
path before it becomes proceced by IsolatedStorageFile and determine if
exists.

Two private methods were added to handle this, it won't affect the
regular workflow if the path is absolute and it will combine it as
usual.