[CB-7419][WP8] Added support to get metada from dir
The error CODE 1, in the callback it was about the a failed attempt to
get the size of a directory using IsolatedStorageFileStream, get that
information, throwing a SystemException 'DirectoryNotFoundException'.
Instead IsolatedStorageFileStream to get that information from a
directory, now uses Application.GetResourceStream, to that information
from the provided path, only if directory. If other file it would use
'IsolatedStorageFileStream'