[CB-7375][Entry] get proper filesystem in Entry
When the native side it doesn't return any kind of value related
directly with the fileSystem, it has to be recreated after a copy,
however, it makes perfect sense that, when the entry for the copied or
moved file if there's no value in entry.fileSystem, it takes the
filesystem property from the parent directory that it has been moved to
or copied to. So instead of return a null object or take the previous
filesystem(origin) from whatever location it was copied from. It would
be better take the parent filesystem name and recreate it.