CB-7445 [BlackBerry10] resolveLocalFileSystemURI - change DEFAULT_SIZE to MAX_SIZE
Setting the quota field has no effect other than to limit the amount of data that
may be written to the filesystem. There are no system limits nor are there any
prompts when an application requests a large quota. Unless the user specifically
requests a lower quota, it will be set to MAX_SIZE by default which provides
behaviour similar to the other Cordova platforms.