fix(types): Adjust types to fix incompatibilities with TypeScript 4.4+
Changes in libdom by the TypeScript project led to conflicting definitions with this plugin.

The key point that joins libdom and this plugin together are via the `root` property
on the FileSystem interface. This results in the DirectoryEntry and Entry types needing
to align with libdom's FileSystemDirectoryEntry & FileSystemEntry.

This introduces 3 slight irregularities in order to be compatible with libdom

1. getParent successCallback needs to be optional
2. getParent errorCallback needs to be typed as a DOMException, despite it being a FileError in reality
3. getFile and getDirectory success types need to be relaxed to the base Entry